% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtradescape.R
\name{buildtradescape}
\alias{buildtradescape}
\title{Hierarchical Bayesian modeling of local adaptation to multidimensional environmental gradients}
\usage{
buildtradescape(
  allele_count,
  snp_idx,
  site_idx,
  envdata,
  coords,
  ploidy,
  cores = 2,
  chains = 2,
  iter = 2000,
  warmup = 1000,
  adapt_delta = 0.9
)
}
\arguments{
\item{allele_count}{Integer vector, in which each element represents the count of the reference allele for one individual at one SNP.}

\item{snp_idx}{Integer vector of same length as allele_count, serving as an index that indicates which SNP each value of allele_count pertains to.}

\item{site_idx}{Integer vector of same length as allele_count, serving as an index that indicates which geographic sampling site each value of allele_count pertains to.}

\item{envdata}{Matrix of real values, in which rows represent individual organisms--not sampling sites--and columns represent environmental variables to be used as predictors of allele frequencies.}

\item{coords}{Matrix of real values, in which rows represent geographic sampling sites, and columns represent x and y geographic coordinates.}

\item{ploidy}{Single integer value representing the ploidy of the individuals in the data set--i.e., the maximum possible value for each element in allele_count.}

\item{cores}{Number of processor cores to use for parallel Hamiltonian Monte Carlo (HMC) posterio sampling.}

\item{chains}{Number of sampler chains for Stan's Hamiltonian Monte Carlo (HMC) sampler.}

\item{iter}{Total number of posterior sampling iterations for Stan's Hamiltonian Monte Carlo (HMC) sampler, including warmup.}

\item{warmup}{Number of warmup iterations for Stan's Hamiltonian Monte Carlo (HMC) sampler.}

\item{adapt_delta}{A parameter that adjusts the behavior of Stan's Hamiltonian Monte Carlo (HMC) posterior sampling algorithm. Increasing closer to 1 can help avoid divergent transitions.}
}
\value{
A stan model object; check back soon for more information.
}
\description{
Hierarchical Bayesian modeling of local adaptation to multidimensional environmental gradients
}
\details{
In development; check back soon. See README at github.com/kyle-rosenblad/tradescape.
}
\examples{

data(ex_data)

# Quick demo run. Will prompt warnings about insufficient posterior sampling.
mod <- buildtradescape(allele_count=ex_data$allele_count,
                      snp_idx=ex_data$snp_idx,
                      site_idx=ex_data$site_idx,
                      envdata=ex_data$envdata,
                      coords=ex_data$coords,
                      ploidy=ex_data$ploidy,
                      warmup=100,
                      iter=200)

}
