% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_dataR.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{A simulated landscape genomic data set}
\format{
An object of class \code{list} of length 6.
}
\usage{
ex_data
}
\description{
List containing:
allele_count, an integer vector of counts of the reference allele for simulated individuals
snp_idx, an integer vector indicating which SNP (i.e., point in the genome) each allele count pertains to
site_idx, an integer vector indicating which geographic sampling site each allele count pertaints to
envdata, a two-column data frame, in which rows are individuals (not sites), and columns are environmental variables describing the conditions at the site where each individual was sampled
coords, a data frame in which rows represent geographic sites and the two columns are x and y coordinates
ploidy, a single integer value indicating the ploidy of the organisms; i.e., the maximum possible allele count
}
\keyword{datasets}
