% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradescape-package.R
\docType{_PACKAGE}
\name{tradescape-package}
\alias{tradescape-package}
\alias{tradescape}
\title{The 'tradescape' package.}
\description{
Hierarchical Bayesian modeling of local adaptation to multidimensional environmental gradients with SNP data
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
